Move mapping parsing to a dedicated file.
